Ignore several auto-generated preview output
Emacs + Auctex generates many different types of files for previewing
documents. We ignore these.

Note that we will not ignore .pdf output because, once the document is
complete we will commit it to the repository so that it is easier for
people to view the document without having to build it from scratch.